Replace Timestring dependency with a newer Fork
This commit replaces Timestring with a newer version that is more actively maintained than Steve Peak's original repository. The alternative to this approach would be to bring in a fork of timestring and update it ourselves. For as minimally as this depedency needs to be updated, though, I'm not entirely certain if this is necessary.